refactor(types): move StopServiceState to types/app/transit
- StopServiceState and StopServiceStateInput were defined in
  domain/transit/timetable-utils, causing a cross-layer circular type
  dependency (types/app/repository and types/app/transit-composed
  importing from domain/).
- Move both types to src/types/app/transit where stable app types live.
- Update timetable-utils to import them from types/app/transit.
- Update all import sites (repository.ts, transit-composed.ts, components,
  stories, tests).
- Fix ROUTE_TYPE_PRIORITY comparator in bottom-sheet.tsx to use
  ?? Number.POSITIVE_INFINITY so unknown route types always sort last
  without NaN.
